# api/openapi.yaml
openapi: 3.0.3
info:
  title: Morning Night Dream - AppGateway
  description: |-
    This is the AppGateway API documentation.
  termsOfService: https://example.com
  contact:
    email: morning.night.dream@example.com
  license:
    name: MIT
  version: 0.0.1
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: http://localhost:8082/api
tags:
  - name: article
  - name: health
paths:
  /v1/article:
    get:
      tags:
        - article
      summary: List articles
      description: List articles
      operationId: v1ListArticles
      parameters:
        - name: pageToken
          in: query
          description: トークン
          required: false
          explode: true
          schema:
            type: string
        - name: maxPageSize
          in: query
          description: ページサイズ
          required: true
          explode: true
          schema:
            type: integer
      responses:
        200:
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListArticleResponse'
        500:
          description: サーバーエラー
  /v1/health:
    get:
      tags:
        - health
      summary: ヘルスチェック
      description: ヘルスチェック
      operationId: v1Health
      responses:
        200:
          description: OK
        500:
          description: Internal Server Error
components:
  schemas:
    Article:
      type: object
      properties:
        id:
          type: string
          description: id
          format: uuid
        url:
          type: string
          description: 記事のURL
          format: uri
          example: https://example.com
        title:
          type: string
          description: タイトル
          example: sample title
        description:
          type: string
          description: description
          example: sample description
        thumbnail:
          type: string
          description: サムネイルのURL
          format: uri
          example: https://example.com
        tags:
          type: array
          description: タグ
          items:
            type: string
            example: tag
    ListArticleResponse:
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: '#/components/schemas/Article'
        nextPageToken:
          type: string
          description: 次回リクエスト時に指定するページトークン